kafka: split out protocol messages type list
It will be shared with tests.
